requirements.txt
fastapi
uvicorn[standard]
redis
types-redis
python-jose[cryptography]
pymongo
motor
starlette
jinja2
python-multipart
bcrypt==2.4.1
kafka-python
email-validator
six
wait-for-it  
python-dotenv
datetime
passlib
